Project/FAQsBot/requirements.txt
streamlit>=1.22.0,<2.0.0
langchain>=0.0.284,<0.1.0
langchain-community>=0.0.10,<0.1.0
langchain-google-genai>=0.0.5,<0.1.0
python-dotenv>=1.0.0,<2.0.0
tiktoken>=0.4.0,<0.6.0
faiss-cpu>=1.7.4,<2.0.0
protobuf>=3.19.0,<5.0.0
numpy==1.26.4
